+++ pyntel4004/test/06_io_ram/test_0200_227_instruction_wpm.py
@@ -185,3 +185,161 @@ def test_wpm_scenario1_write(rambank, chip, register, address):
 
     # Pickling each chip and comparing will show equality or not.
     assert pickle.dumps(chip_test) == pickle.dumps(chip_base)
+
+
+@pytest.mark.parametrize("rambank", [0, 7])
+@pytest.mark.parametrize("chip", [2, 3])
+@pytest.mark.parametrize("register", [1, 3])
+@pytest.mark.parametrize("address", [2, 5, 3])
+def test_wpm_scenario1_read(rambank, chip, register, address):
+    """Test WPM instruction functionality."""
+
+    chip_test = processor()
+    chip_base = processor()
+
+    '''
+    This piece of code reads an 8-bit value (consisting of 2 4-bit values)
+    from the program ram in 2 chunks. The setting up of this code and the
+    execution of several WPM statements is key to the testing, although much
+    has to be set up around those instructions to stage the tests.
+
+    1     FIM        0P    0
+    2     SRC        0P          / Select DATA RAM chip 0 register 0.
+    3     RD1                    / Read middle 4 bits of address.
+    4     XCH        10          / Save in register 10.
+    5     RD2                    / Read lowest 4 bits of address.
+    6     XCH        11          / Save in register 11.
+    7     RD0                    / Read highest 4 bits of address.
+    8     FIM        0P    240
+    9     SRC        0P          / Select ROM port 15.
+    10    WRR                    / Write high address.
+    11    SRC        5P          / Write middle + low address (RP5)
+
+    12    WPM                    / PRAM data to ROM port 14.
+    13    WPM                    / PRAM data to ROM port 15.
+
+    14    FIM        0P    224
+    15    SRC        0P          / Select ROM port 14.
+    16    RDR                    / Read to accumulator
+    17    XCH        2           / Save in register 2
+    18    INC        0
+    19    SRC        0P          / Select ROM port 15
+    20    RDR                    / Read to accumulator
+    21    XCH        3           / Save in register 2
+    '''
+    # Perform the instruction under test:
+    # Preamble.....
+    chip_test.PROGRAM_COUNTER = 0             # Set PC to be zero
+    chip_test.CURRENT_RAM_BANK = rambank      # Set RAM BANK to be the supplied
+
+    chip_base.PROGRAM_COUNTER = 0             # Set PC to be zero
+    chip_base.CURRENT_RAM_BANK = rambank      # Set RAM BANK to be the supplied
+
+    register_pair = 5
+    reg_pair_first = (register_pair * 2)
+    reg_pair_second = reg_pair_first + 1
+
+    # Lines 1 - 2
+    chip_test.COMMAND_REGISTER = \
+        encode_command_register(chip, register, address, 'DATA_RAM_CHAR')
+
+    address_to_write_to = convert_to_absolute_address(
+        chip_test, rambank, chip, register, address)
+    chunks = c2n(12, 4, address_to_write_to, 'b')
+
+    # Lines 3 - 4    # Store middle bits in register "reg_pair_first"
+    chip_test.STATUS_CHARACTERS[rambank][0][0][1] = \
+        binary_to_decimal(str(chunks[1]))
+    chip_test.REGISTERS[reg_pair_first] = \
+        chip_test.STATUS_CHARACTERS[rambank][0][0][1]
+
+    # Lines 5 - 6   # Store lower bits in register "reg_pair_second"
+    chip_test.STATUS_CHARACTERS[rambank][0][0][2] = \
+        binary_to_decimal(str(chunks[2]))
+    chip_test.REGISTERS[reg_pair_second] = \
+        chip_test.STATUS_CHARACTERS[rambank][0][0][2]
+
+    # Lines 7 - 10   # Store higher bits in ROM PORT 15
+    chip_test.STATUS_CHARACTERS[rambank][0][0][0] = \
+        binary_to_decimal(str(chunks[0]))
+    chip_test.ROM_PORT[15] = chip_test.STATUS_CHARACTERS[rambank][0][0][0]
+
+    # Line 11
+    chip_test.COMMAND_REGISTER = processor.read_registerpair(chip_test, 5)
+
+    # Lines 12 - 13
+    processor.wpm(chip_test)
+    processor.wpm(chip_test)
+
+    # Lines 14 - 16
+    # Get the value from ROM port 14 to accumulator
+    chip_test.set_accumulator(chip_test.ROM_PORT[14])
+
+    # Line 17
+    chip_test.REGISTERS[reg_pair_first] == chip_test.read_accumulator()
+
+    # Lines 18 - 20
+    # Get the value from ROM port 15 to accumulator
+    chip_test.set_accumulator(chip_test.ROM_PORT[15])
+
+    # Line 21
+    chip_test.REGISTERS[reg_pair_second] == chip_test.read_accumulator()
+
+    # Simulate conditions in base chip
+    # Lines 1 - 2
+    chip_base.COMMAND_REGISTER = \
+        encode_command_register(chip, register, address, 'DATA_RAM_CHAR')
+
+    address_to_write_to = convert_to_absolute_address(
+        chip_base, rambank, chip, register, address)
+    chunks = c2n(12, 4, address_to_write_to, 'b')
+
+    # Lines 3 - 4    # Store middle bits in register "reg_pair_first"
+    chip_base.STATUS_CHARACTERS[rambank][0][0][1] = \
+        binary_to_decimal(str(chunks[1]))
+    chip_base.REGISTERS[reg_pair_first] = \
+        chip_base.STATUS_CHARACTERS[rambank][0][0][1]
+
+    # Lines 5 - 6   # Store lower bits in register "reg_pair_second"
+    chip_base.STATUS_CHARACTERS[rambank][0][0][2] = \
+        binary_to_decimal(str(chunks[2]))
+    chip_base.REGISTERS[reg_pair_second] = \
+        chip_base.STATUS_CHARACTERS[rambank][0][0][2]
+
+    # Lines 7 - 10   # Store higher bits in ROM PORT 15
+    chip_base.STATUS_CHARACTERS[rambank][0][0][0] = \
+        binary_to_decimal(str(chunks[0]))
+    chip_base.ROM_PORT[15] = chip_base.STATUS_CHARACTERS[rambank][0][0][0]
+
+    # Line 11
+    chip_base.COMMAND_REGISTER = processor.read_registerpair(chip_test, 5)
+
+    # Lines 12 - 13
+    chip_base.ROM_PORT[14] = chip_base.PRAM[address_to_write_to] >> 4 << 4
+    processor.flip_wpm_counter(chip_base)
+    chip_base.increment_pc(1)
+    chip_base.ROM_PORT[15] = chip_base.PRAM[address_to_write_to] << 4 >> 4
+    processor.flip_wpm_counter(chip_base)
+    chip_base.increment_pc(1)
+
+    # Lines 14 - 16
+    # Get the value from ROM port 14 to accumulator
+    chip_base.set_accumulator(chip_base.ROM_PORT[14])
+
+    # Line 17
+    chip_base.REGISTERS[reg_pair_first] == chip_base.read_accumulator()
+
+    # Lines 18 - 20
+    # Get the value from ROM port 15 to accumulator
+    chip_base.set_accumulator(chip_base.ROM_PORT[15])
+
+    # Line 21
+    chip_base.REGISTERS[reg_pair_second] == chip_base.read_accumulator()
+
+    # Make assertions that the base chip is now at the same state as
+    # the test chip which has been operated on by the instruction under test.
+    assert chip_test.read_program_counter() == chip_base.read_program_counter()
+    assert chip_test.read_accumulator() == chip_base.read_accumulator()
+
+    # Pickling each chip and comparing will show equality or not.
+    assert pickle.dumps(chip_test) == pickle.dumps(chip_base)